Big checkin.
1. Refactored the ArtworkInfo data model.

2. Implemented proper image blurring. Nextjs uses
base64 encodings that need to be either statically
imported or passed in with the plaiceholder library.

Most of the checkin was a significant refactor related to
enabling this new plaiceholder functionality for the image
blurring. Since this is an art website, images are a
really important factor and I felt it was critical to
nail this optimization.