wip of my first attempt to implement a firebase admin based backend.
I have an issue where the obtained signed URL links to
production google servers rather than the emulated storage,
and calling getDownloadURL to obtain the public url just
straight up crashes the emulator.

Some relevant links to investigate that particular issue.
https://github.com/firebase/firebase-admin-node/issues/1352
https://github.com/firebase/firebase-tools/issues/3396
https://github.com/invertase/react-native-firebase/issues/6638
https://stackoverflow.com/questions/42956250/get-download-url
-from-file-uploaded-with-cloud-functions-for-firebase

This changelist will reproduce the issue if you try to change to
getDownloadURL or try to get an emulated storage url

THIS IS NOT A FUNCTIONAL CHANGELIST.